Use cublasStatus_t instead of cudaError_t when error checking cublas calls
Cublas success is indicated by CUBLAS_STATUS_SUCCESS. This must be used
to determine if a cublas function call was successful.

Although Cuda does not complain if we use cudaError_t, when bulding with
ROCm support it is absolutely necessary to use the correct enum
(cublasStatus_t), else ROCm will always interpret the returned
status as an error and crash.